seed/requirements.txt
psycopg2-binary==2.9.*
pymongo==4.6.*
faker==22.*
